feat: simpa as an elab
There were a bunch of small issues in `simpa` as a result of it being
implemented essentially as a macro. By implementing it directly, we can
give much better error messages in the edge cases.

Also, the "try `simp`" error message has been downgraded to a
linter warning, enabled by default. It isn't actually using the linter
framework, which has the usual downside that you might get a linter
warning for `split <;> simpa` if `simpa` can be simplified to `simp`
in only one branch.